requirements.txt
requests>=2.25.1
pyjwt>=2.0.0
python-dotenv>=0.19.0 